requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.1
rank-bm25>=0.2.2
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
gradio>=4.44.0
click>=8.1.7
rich>=13.7.0
pyyaml>=6.0.1
elasticsearch>=8.13.0
python-dotenv>=1.0.1
tqdm>=4.66.4
